Replace implicit conversions with toService method
