+++ benchmarks/src/main/scala/io/finch/benchmarks/service/service.scala
@@ -44,12 +44,13 @@ class FinchUserService(implicit
   val deleteUsers: Service[HttpRequest, HttpResponse] =
     Service.const(db.delete().map(count => Ok(s"$count users deleted")))
 
-  val users: Service[HttpRequest, HttpResponse] =
+  val users: Service[HttpRequest, HttpResponse] = (
     Get    / "users" / long /> getUser :+:
     Get    / "users" /> allUsers       :+:
     Post   / "users" /> createUser     :+:
     Put    / "users" /> updateUser     :+:
     Delete / "users" /> deleteUsers
+  ).toService
 
   val handleExceptions = new SimpleFilter[HttpRequest, HttpResponse] {
     def apply(req: HttpRequest, service: Service[HttpRequest, HttpResponse]): Future[HttpResponse] =
