+++ demo/src/main/scala/io/finch/demo/Main.scala
@@ -100,7 +100,7 @@ object Demo {
 
   // An API endpoint.
   val api: Service[AuthRequest, HttpResponse] =
-    getUser :+: getUsers :+: postUser :+: postTicket
+    (getUser :+: getUsers :+: postUser :+: postTicket).toService
 
   // An HTTP endpoint with exception handler and Auth filter.
   val backend: Service[HttpRequest, HttpResponse] =
